requirements.txt
flask
pandas
xhtml2pdf
openai
xhtml2pdf
python-dotenv